refactor: ディレクトリ名をpagesに変更
分かりやすい名前にするため
